+++ temba/channels/types/jasmin/views.py
@@ -1,7 +1,7 @@
 import phonenumbers
 
 from django import forms
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from temba.channels.views import ALL_COUNTRIES, AuthenticatedExternalCallbackClaimView, ClaimViewMixin
 from temba.utils.fields import ExternalURLField, SelectWidget