+++ temba/channels/types/viber_public/type.py
@@ -1,7 +1,7 @@
 import requests
 
 from django.urls import reverse
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from temba.contacts.models import URN