+++ temba/orgs/views.py
@@ -8,7 +8,7 @@ from datetime import datetime, timedelta
 from decimal import Decimal
 from email.utils import parseaddr
 from functools import cmp_to_key
-from urllib.parse import parse_qs, unquote, urlparse
+from urllib.parse import parse_qs, quote, unquote, urlparse
 
 import iso8601
 import pyotp
@@ -47,12 +47,11 @@ from django.http import HttpResponse, HttpResponseRedirect, JsonResponse
 from django.shortcuts import resolve_url
 from django.urls import reverse, reverse_lazy
 from django.utils import timezone
-from django.utils.encoding import DjangoUnicodeDecodeError, force_text
+from django.utils.encoding import DjangoUnicodeDecodeError, force_str
 from django.utils.html import escape
-from django.utils.http import urlquote
 from django.utils.safestring import mark_safe
 from django.utils.text import slugify
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import View
 
@@ -221,7 +220,7 @@ class ModalMixin(SmartFormView):
         if "success_url" in kwargs:  # pragma: no cover
             context["success_url"] = kwargs["success_url"]
 
-        pairs = [urlquote(k) + "=" + urlquote(v) for k, v in self.request.GET.items() if k != "_"]
+        pairs = [quote(k) + "=" + quote(v) for k, v in self.request.GET.items() if k != "_"]
         context["action_url"] = self.request.path + "?" + ("&".join(pairs))
 
         return context
@@ -498,7 +497,7 @@ class LoginView(Login):
             verify_url = reverse("users.two_factor_verify")
             redirect_url = self.get_redirect_url()
             if redirect_url:
-                verify_url += f"?{self.redirect_field_name}={urlquote(redirect_url)}"
+                verify_url += f"?{self.redirect_field_name}={quote(redirect_url)}"
 
             return HttpResponseRedirect(verify_url)
 
@@ -1282,7 +1281,7 @@ class OrgCRUDL(SmartCRUDL):
                 # check that it isn't too old
                 data = self.cleaned_data["import_file"].read()
                 try:
-                    json_data = json.loads(force_text(data))
+                    json_data = json.loads(force_str(data))
                 except (DjangoUnicodeDecodeError, ValueError):
                     raise ValidationError(_("This file is not a valid flow definition file."))
 
@@ -2833,7 +2832,7 @@ class OrgCRUDL(SmartCRUDL):
                 surveyors_group = Group.objects.get(name="Surveyors")
                 token = APIToken.get_or_create(org, user, role=surveyors_group)
 
-                org_name = urlquote(org.name)
+                org_name = quote(org.name)
 
                 return HttpResponseRedirect(
                     f"{self.get_success_url()}?org={org_name}&uuid={str(org.uuid)}&token={token}&user={username}"